feat(retrieval): knowledge-fit scoring + efficiency fixes
Knowledge-fit (with eval gate: seeded eval-retrieval MRR 0.885 -> 0.904):
- Strong-signal FTS short-circuit now applies the same metadata scoring as
  fusion via a new shared _metadata_score_factor (type weight x feedback x
  priority x decay) — raw text scores let a lexically-sharp task_completion
  (weight 1.4, 1916 rows) outrank a decision/lesson (2.0) that would have won
  under fusion. No retrieval path can bypass knowledge weighting now.
- task_completion decay half-life 46d -> 23d (episodic exhaust, 40x more
  numerous than decisions; was decaying at the decision rate).
- Access-aware decay slowdown extended from decision-only to lesson_learned,
  advisor_insight, skill_template — recalled knowledge persists longer.
- apply_strength_decay: protect decision/lesson/insight (soft-delete is
  deletion for practical purposes) + single-transaction batch (was one
  lock+commit+WAL checkpoint per decayed row).
- Temporal inference word-boundary fix: bare month/weekday tokens require a
  temporal preposition — "we may need", "the march of progress", and
  "monday-morning quarterbacking" no longer trigger date-range scoring.

Efficiency:
- Failed-import caches for omega.thompson / omega.pattern_learner /
  omega.embedding_client — Python does not cache failed imports and these
  were probed per candidate per query (50-100 sys.path walks per query).
- Batch hydration (one IN query, was N+1 SELECTs): temporal channel,
  query_by_type, find_similar.
- Hot-cache refresh filters on the indexed status column instead of an
  unindexable json_extract full scan.
- Reranker: default precision int8 (fp32 was a 2.3GB implicit download);
  _RERANK_CANDIDATES 10 -> 20 so the position-aware rank-11+ tier is
  reachable (was dead code).
- omega doctor: new check that query expansion has a working LLM provider
  (it silently no-ops without one).

Tests: 15 new TDD tests (failing first) in test_retrieval_knowledge.py + a
strength-decay protection regression; two reranker precision tests updated
to pin the int8 default.

+++ src/omega/reranker.py
@@ -5,7 +5,7 @@ Provides:
 - cross_encoder_score(query, passages) → list of relevance scores (or None)
 - Lazy model loading with circuit breaker (3 attempts, 5-min cooldown)
 - Env-var disable: OMEGA_CROSS_ENCODER=0
-- Precision selection: OMEGA_RERANKER_PRECISION=fp32|int8 (default: fp32)
+- Precision selection: OMEGA_RERANKER_PRECISION=fp32|int8 (default: int8)
   fp32: full precision, ~2.3 GB on disk, ~2.5 GB RSS
   int8: quantized, ~571 MB on disk, ~650 MB RSS
 - Auto-download on first use: OMEGA_RERANKER_AUTODOWNLOAD=1 (default).
@@ -51,7 +51,11 @@ _RERANKER_MODEL = None  # Tuple of (tokenizer, session) when loaded
 _AVAILABLE_MODELS = {
     "bge-reranker-v2-m3": {
         "repo_id": "onnx-community/bge-reranker-v2-m3-ONNX",
-        "default_precision": "fp32",
+        # int8 default (fork): fp32 is a 2.3 GB download + ~2.5 GB RSS per
+        # process — unreasonable as an implicit default for a single-user
+        # store. int8 is ~571 MB / ~650 MB RSS with near-par quality.
+        # Set OMEGA_RERANKER_PRECISION=fp32 to opt back in.
+        "default_precision": "int8",
         "precisions": {
             "fp32": {
                 "dir": "~/.cache/omega/models/bge-reranker-v2-m3-onnx",
